+++ libs/config/src/lib/config.service.ts
@@ -39,6 +39,29 @@ export class ConfigService<Config extends Record<string, any> = Record<string, a
 
   public static Urls = [ 'config.json' ];
 
+  public static async SideLoad(url: string, propertyPath: string, required?: boolean): Promise<void> {
+
+    if (!this.Config) {
+      throw new Error('Config side load is only possible after the initial config load.');
+    }
+
+    try {
+
+      const response = await fetch(url);
+      SetObjectValue(this.Config, propertyPath, await response.json());
+
+    } catch (error) {
+      if (required) {
+        throw new Error(`Could not side load config '${url}': ${error.message}`);
+      } else {
+        console.warn(url, error.message);
+      }
+    }
+
+    console.debug(`app config side load '${propertyPath}'`, this.Config);
+
+  }
+
   /**
    * Used to load the app config from a remote resource.
    *