+++ libs/forms/src/lib/directives/form.directive.ts
@@ -6,7 +6,6 @@ import {
   OnInit,
   EventEmitter,
   Output,
-  HostListener,
   HostBinding,
   ChangeDetectorRef,
   Input,
@@ -286,7 +285,6 @@ export class FormDirective<T extends Record<string, any> = any>
     }
   }
 
-  @HostListener('submit', ['$event'])
   public onSubmit($event: Event): boolean {
     $event.preventDefault();
     super.onSubmit($event);
@@ -410,13 +408,15 @@ export class FormDirective<T extends Record<string, any> = any>
   }
 
   protected getSubmitValue(): T {
-    let value: T;
+    let value: T = undefined as any;
 
     if (typeof this._formDefinition[ 'getSubmitValue' ] === 'function') {
       value = this._formDefinition.getSubmitValue();
-    } else {
-      value = this.form.value;
+    } else if (typeof this._formDefinition[ 'toJSON' ] === 'function') {
+      value = this._formDefinition.toJSON();
     }
+
+    value = value ?? this.form.value;
 
     return clone(value);
   }